updated SerializedTransaction
