+++ zaino-fetch/src/primitives.rs
@@ -4,7 +4,7 @@
 
 // pub mod address;
 // pub mod block;
-pub mod chain;
+// pub mod chain;
 pub mod error;
 pub mod height;
 pub mod transaction;
